feat: separated memory and rag, will work later
